+++ tests/test_neutron_animation_controller.py
@@ -1,5 +1,5 @@
 from PySide2.QtGui import QMatrix4x4
-from nexus_constructor.NeutronAnimationController import NeutronAnimationController
+from nexus_constructor.neutron_animation_controller import NeutronAnimationController
 from mock import Mock
